% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.curve.R
\docType{data}
\name{points.curve}
\alias{points.curve}
\title{points.curve function to calculate the coordinates (fpr,sen)}
\format{An object of class \code{NULL} of length 0.}
\usage{
points.curve}
\arguments{
\item{xsample}{sample etiquetes}

\item{ysample}{vector with the values of sensibility of a model}
}
\value{
return a matrix with a serie of coordenates that will be used to generate a ROC curve
}
\description{
Function that calculates a serie of coordinates that will be used to do the curve ROC.
This function sorts a vector resulted of a model test and generates several points that will be used to divide the ROC curve and analyse it
}
\examples{
using examp
le data from our package
data("intersticial")
intersticial <- intersiticial
sample <- sample(intersticial, dim(intersticial[, 1])/10)
dataset.boot<-intersticial[sample, ]
setsample<-dataset.boot[, 1]
datasample<-dataset.boot[, 2]
points <- points.curve(setsample, datasample)
}
\keyword{datasets}
