% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TpAUC.R
\name{TpAUC}
\alias{TpAUC}
\title{Tigher partial area under the ROC curve}
\usage{
TpAUC(
  dataset,
  low.value = NULL,
  up.value = NULL,
  plot = FALSE,
  selection = NULL,
  variable = NULL
)
}
\arguments{
\item{dataset}{Dataframe of the complete information of the samples}

\item{low.value}{lower false positive rate value that the function will use to calculate the pAUC}

\item{up.value}{upper false positive rate value that the function will use to calculate the pAUC}

\item{plot}{ROC plot}

\item{selection}{vector that will only be used if the parameter "dataset" is a RangedSummarizedExperiment object.
This parameter is used to select the variables that will be analysed}

\item{variable}{in case that dataset is a  SummarizedExperiment, indicate the Gold Standard}
}
\value{
RangedSummarizedExperiment object with the pAUC and the TpAUC scores,and the TPR and FPR values for each ROC curve generated
}
\description{
It standarizes the partial area under the ROC curve by the tigher index
}
\examples{
library(fission)
data("fission")
resultsT <- TpAUC(fission, low.value = 0, up.value = 0.25, plot = TRUE, selection = c("SPNCRNA.1080","SPAC186.08c"), varidable="strain")
}
